VMS-11002: UpdateManager::start() logic changed a bit
<==== Mercurial data ====>
Branch: x_vms_11002_nx_update_refactoring

+++ common_libs/nx_update/src/nx/update/common_update_manager.cpp
@@ -42,14 +42,18 @@ UpdateStatus CommonUpdateManager::start()
     nx::update::Package package;
     UpdateStatus updateStatus;
 
-    m_downloaderFailed = false;
-    for (const auto& file : downloader()->files())
+    bool shouldDownload = statusAppropriateForDownload(&package, &updateStatus);
+    if (shouldDownload || !package.isValid())
     {
-        if (file.startsWith("update/"))
-            downloader()->deleteFile(file);
+        m_downloaderFailed = false;
+        for (const auto& file : downloader()->files())
+        {
+            if (file.startsWith("update/"))
+                downloader()->deleteFile(file);
+        }
     }
 
-    if (!statusAppropriateForDownload(&package, &updateStatus))
+    if (!shouldDownload)
         return updateStatus;
 
     using namespace vms::common::p2p::downloader;
